The curried role can be bound
As far as Rakudo is concerned, a role binding is just the role candidate
we already check for during the completion of a parameterization, but is
given a more formal, generic interface. In defining the logic for
producing a candidate in the role group, we eventually gain the ability
to cancel a completion.

Metamodel::CurriedRoleHOW's complete_parameterization method must depend
on bindings immediately, as builds take an insufferable amount of time
with the extra attribute loafing around. In doing so, we allow the
following to be valid:

role Foo[Any ::T] { }
role Bar[Cool ::T] does Foo[T] { }