"Hide" use of boot-resume for proto onlystar
We use dispatch resumption for the {*} in a proto with a body. This
should not count as a resumable dispatch in a user-visible sense,
however. Thus make sure if we are asked to resume and cannot in this
context, then we report an error, rather than silently giving a Nil.