Flatten the role/role typecheck
We need to invert the role/role typecheck, but this results in an ugly
recursive typecheck that should result in strict equality checks for
doees anyways. Because we now do a literal typecheck on the role being
curried, we need to consider all parents of the role being curried.

This makes the Blob[uint8]/Blob[uint8] typecheck ~2x faster.