Make .sort aware of any &[cmp] additions
This makes .sort aware of any local &[cmp] additions from where .sort
is called, if called without an explicit comparator.

This will use CALLERS:: to find out if the outer &[cmp] is different
from the "system" &[cmp], and will use that if that is the case.

The use of CALLERS:: is a bit icky, and I'm not sure that this workaround
will be needed as soon as new-disp lands.

Also took the opportunity to reduce the amount of nqp code in this logic.

Has no discernable performance effect on any sorting, except on sorting
empty lists.