Fix argument to nqp::hllize inadvertently getting a name
The value of a Pair in an arglist gets a .named annotation. The p6callmethodhow
desugar re-used the existing QAST::Op node, modified it and used it as argument
to nqp::hllize. The new dispatcher based implementation doesn't cope with the
name though. Fix by creating a new un-named QAST::Op and copying over the data
we need.