Introduce -Msafe-snapper
Snapper functionality that allows you to press Control-c and *still*
get a Telemetry report.  Removes that functionality from -Msnapper.

Problem with Control-c safety is that it will start up the supervisor
thread to handle the pressing of Control-c, which causes extra load
and reduces the number of available threads, thus affecting the Telemetry
data.