Fix false positive in basic telemetry test
It's possible (especially with ROAST_TIMING_SCALE) that Telemetry takes
multiple snapshots while we give it time to do at least one. So adjust the
test to expect at least one snapshot to be taken, but be fine with moar than
one. This fits the test description of "did the snapper start taking snaps"