Allow nextsame et al in detached method calls
That is, when we look up a method and then separately invoke it. We
supported this historically, so keep it working. This also fixes the
case where we have an NQP-based type with a Raku method installed in it
that then does nextsame, which shows up in various modules that built
upon the MOP, such as Method::Also (which is fixed by this change).