Don't run every proto during CORE.setting build
We can't properly do the analysis as to whether or not we need to run
the proto since we're not yet properly bootstrapped. However, assume
that we don't need to run any of them. This saves rather a lot of time,
because to run the proto involves compiling it. It also at present
triggers the compilation of all multi candidates and would compile them
in duplicate.