Remove use of Str.new from grammar
FROGGS added this in 2014, when perhaps this was needed because of
Parrot.  Removing it completely and just using whatever self.orig
gives us, does not break any spectest, nor does it inhibit installing
slangs like Slang::Tuxic or Slang::Piersing.

So I assume this was really an old wart, which is now gone.  Should
a blin run show that this *did* break something, then just revert this.