Fix escaping of # in regexes
In the MAIN language, `foo \# ` is correctly parsed as `foo` followed
by an unspace and then a comment.  In the regex Slang, however,
unspaces are not allowed and that should be parsed the same as `foo
'#'`.

However, a 2013 commit to add support for embedded comments in regexes
inadvertently allowed unspaces as well, leading to misparsing of \# in
some cases.  This commit changes the implementation of the 2013 commit
to allow embedded comments without incorrectly allowing unspaces.