Extend pseudos coverage into more namespaces
Static chains now scan `$?PACKAGE` (`OUR`) on top of lexical outers.

Dynamic chains try to include `$!dynamic_context` of `$*PROMISE`, and
`GLOBAL` and `PROCESS` too. For the latter two name mapping is
implemented. As it is expected for dynamic pseudos to work with
`*`-twigilled dynamic symbols, all twigil-less symbols on `PROCESS` and
`GLOBAL` get twigils when accessed via a dynamic chain pseudo. I.e.
same way as one uses `$*IN` in the code they can now refer to it as
DYNAMIC::<$*IN>.