Optimize method calls made on roles
We can pre-calculate and cache the pun in the callsite, rewriting the
call capture in the dispatcher to include it. This means that calls to
methods on a role (typically `.new`, which relies on punning) will now
no longer need to slurp/flatten arguments, will avoid an extra level of
invocation, and expose the target method in the pun for spesh linking
and/or inlining.