Add a candidate of Promise.at that accepts a DateTime
Passing a DateTime to Promise.at seems natural and it strikes me that a
common use case will be to pass a date time that one might have at hand.

This adds a candidate with a coercion to Instant for `$at` (so will also
accept a custom class that implements .Instant,) and refines the
existing candidate to coerce to Numeric() so as to make the error on
being passed something that can't be converted to Numeric clearer.