Fix BEGIN-time calls to bind-check multis
When we have multi candidates with `where` clauses, unpacks, or named
arg type checks, we invoke them such that a bind failure will cause a
multiple dispatch resumption. If a multi is called at compile time and
uses these features, we need to make sure that the candidate is compiled
in advance of ever invoking it; otherwise, the thing we invoke is really
a compiler thunk, and the resumption on bind failure falsely ends up
applying to it, and is thus hidden from the compiled multi.