Expose more constants in raku-invoke dispatcher
When the code object we're invoking is a known constant in the dispatch
- typically because we got it out of the methods table - we don't need
the dispatch program to look up the $!do dynamically. We can instead
just look it up directly and expose it to the VM as a constant. This
results in shorter dispatch programs and lets spesh avoid inserting
lots of static frame guards.