Remove Scalar containers caller side when possible
Add a bit field to Signature that indicates which arguments are
readonly. Use it in the dispatcher in order to remove values from
Scalar containers on the caller side, rather than leaving it to the
callee. See the large comment inside the commit explaining the multiple
advantages this provides.

There cost of doing this is that we run the risk of .VAR getting
unwrapped both by this *and* in the callee. To avoid this, have .VAR do
the wrapping inside of a special ScalarVAR.