Ensure Pair keys remain uncontainerized
A change in 00bb02935 to take advantage of caller-side decont caused a
change in semantics: the Cool type forces a defensive wrapper container
(due to Iterable ~~ Cool), and the removed `nqp::decont` thus meant the
key was stored containerized. Switch the type constraint to Str (which
is the overwhelmingly common case of Pair construction) in order to
reinstate existing behavior, which is depended on in modules.