Add a missing guard in assignment dispatcher
The slower-path assignment with non-nominal type check implementation
depends on getting a Scalar with a simple container descriptor, not one
that is, for example, carrying an array auto-vivification. However, a
guard was not put in place, meaning that the simple container descriptor
implementation could be used on an auto-viv one, resulting in a spurious
type check failure.