Re-implement signature constraint support for parameters
A major change this commit introduces is relocation of signature
object from inside a `where` block into an attribute of
`Parameter::SignatureConstraint` role which is mixed into `Parameter`
when necessary. Correspondingly, constraint is now checked as part of
signature binding when and only when the signature is used in parameter
declaration.

This commit also introduces support for generic signature constraints.
I.e.:

    sub foo(::T, :&fn:(T)) {...}

now works as expected.

1 TODO in t/spec/S03-smartmatch/signature-signature.t is unlocked,
but t/spec/S06-signature/closure-parameters.t is broken.