Fix suggested method ordering on not found error
As moon-child++ noted, "42".matc( did **not** yield 'match' as an
alternative.  This was caused by the ordering of the suggestions
to be on the actual (lowercase) string, rather than at the actual
distance.  And then limiting them to the first 4.

This commit fixes the ordering, and makes sure that *all* suggestions
with the same score as the 4th suggestions, will also be shown.

Fixes #4552