Make sure suggestions are sorted consistently
Before, they were (erroneously) sorted on the lowercase string value
only.  0152e3c38303d7239 fixed that, but introduced an inderminate
ordering on suggestions with the same score.  This commit re-establishes
the old behaviour with regard to suggestions that have the same score.

Fixes a flapper in t/05-messages/01-errors.t line 230, nine++ for spotting