Add guard for container descriptor $!of for non-nominals
Without this guard under certain conditions it is possible that a
container with non-nominal `.of` (like a subset) would take over the
optimization path of assigning `Nil` to a nominal container.

There're two scenarios where this problem occurs. The first is
`nqp::clone` of attribute's `auto_viv_container` and assign to the clone
bound to a lexical.

The second path is to `nqp::create(Scalar)` and set its descriptor to
attribute's container descriptor.

Don't know if there is way to experience this behavior without
involvement of nqp ops.