Get callwith working with complex dispatches
When we combine different kinds of dispatch and also have callwith, the
arguments need to apply to all outer dispatches too. For example, if we
have a wrap of a multi method, and do a `callwith` in the wrapper, then
the arguments have to apply to the multi and method dispatches too. A
description of the strategy can be found in the large comment added in
this commit.