Subtype parameters by their type constraints in Parameter.ACCEPTS
Parameters can be of a type, but not accept all values of that type in a
number of different ways:

- Int $nat where * >= 0
- @tuple [$, $]
- |tuple ($, $)
- &f:($)

i.e. post constraints, subsignatures, and signature constraints.

Parameters do not smartmatch accordingly, but allowing them to do so is
a matter of removing some paranoid else blocks. While we're here, take
care of some nits with the other half of each.