Prevent heredocs continuing after a block
This change addresses proper handling of heredocs
raised in issue #4539 where the docs implied
a heredoc could be defined following a block.

Jonathan said that was not intended and explained
how to fix it. That was the only change to the
code and it had the desired effect: a heredoc
cannot be defined following a block, but can
be used following a function call.

New tests have been added for this change in roast
file 'S02-literals/heredocs.t'.