Make GLOBAL shared at compunit load time
Previously a compunit was installing own GLOBALish as GLOBAL
unconditionally, effectively making it different from loading process'
GLOBAL. This resulted in cases where module's globals are kind of lost
at run time. For example:

    unit module Foo;
    GLOBAL::<$FOO> = 42;
    say $*FOO; # 42
    sub foo is export {
        say $*FOO; # ERROR!
    }

If `foo` is invoked by a script then `$*FOO` is not visible because it
is imported as `$FOO` into script's UNIT namespace and the GLOBAL &foo
sees at run-time is not module's GLOBALish, but scipt's GLOBAL which
lacks the symbol.