Introduce Parameter.untyped
A parameter can lack a type, in which case it will have a type of Mu.
Checking for this alone is not enough to say whether or not a parameter
is untyped; there may be post-constraints, subsignatures, or signature
constraint with which to reject arguments. This is kind of annoying to
introspect from outside Parameter.