Ensure all methods use the updated comp-unit-id (#4578)
In a recent change the distribution id was added to the module name when generating a comp-unit-id for `need` to allow multiple versions of the same name to be loaded. However,`need` and `resolve` need to generate identical comp-unit-ids for a given module name since `resolve` will be called by CUR::PrecompilationRepository when `need`ing a module.