Don't guard Mu-typed params in multi-dispatch
When all candidates encountered have a parameter typed as Mu, then we
do not need to put guards on the argument type. This helps avoid caches
blowing up in size in cases like STORE_AT_KEY, which has a number of
candidates with Mu-typed value parameters that will encounter many
types.