Fix callwith in multi method
More complex examples involving both a method in a base class *and*
a multi method in the subclass worked fine, however when there was
no base class, and a `callwith` to a more general `multi` in the same
class, the more general candidate didn't get invoked. Addresses multiple
of the regressions reported in #4583.