Clarify let and temp operators
Make sure they preserve and restore holes in arrays (though preserving
currently depends on `clone` method implementation) and conterization of
elements.

The new implementation shifts the locus of control to `Array` and `Hash`
themselve, leaving it up to `Rakudo::Internals` to deal with `Scalars`.
The latter was done to allow the backend VM take better care of
optimizations.

With regard to the non-scalar containers, they're expected to provide
two methods: `TEMP-LET-LOCALIZE` and `TEMP-LET-RESTORE`. The first is
expected to return a handle which would let `TEMP-LET-RESTORE` recover
the initial state of the container. The handle object is not constrained
in neither way and it is totally up to the implementation to choose what
the handle is.

The reason for this new approach is that normally a container would know
better how to preserve and restore its internal state. This new approach
also allows for third-party containers to provide full support for `let`
and `temp` operators with ease.