Make sure return value of Str.match is threadsafe
Str.match is supposed to set $/, which currently appears to have
scoping issues, which made it unthreadsafe to e.g. use in a race.map
(see #4601).  Before this commit, the returned value was a
decontainerized version of $/, which had just before been set with
the (non)result of the match.  This commit changes that to return
the match value instead, after having assigned that to $/ in the
outer scope.

I think we seriously need to re-consider $/ in future language
versions of Raku.  It's turning more and more from a DWIM into a
WAT in a threaded environment.