Abstract virtual core META6.json into a module
The install-core-dist.raku script kept the information about which
files are actually part of the Rakudo core distribution to itself.

This commit extracts that information into a separate module, so
that other programs can get at the meta information in a reliable
manner.

This adds a lib/Rakudo/CORE/META.rakumod file that makes the meta
information available in %Rakudo::CORE::META when loaded.

   use Rakudo::CORE::META;
   say "Contains %Rakudo::CORE::META<provides>elems() modules";