Add X::Bind::Rebind error
This adds a Bind::Rebind error that provides more detailed error
message about why the LHS cannot be rebound.  To provide this
info (and to make sure that we don't throw a *re*binding error for
literals or other items that can't be bound in the first place), we
inspect the binding's $target and pass relevant info to the
exception.