Install uninitialized attribute descriptor
We install it as the descriptor in the case that there is an attribute
that has either a default build closure or a required attribute. So far
we don't look at it to determine anything, but the mere act of adding it
does not lead to any test or spectest regressions.