Eliminate nqp::attrinited in clone
It seems like nothing more than an optimization anyway, and even if it
is intended to preserve the partially initialized state of an object, it
will do so under the new scheme too. Also use `nqp::clone_nd` in place
of a `nqp::clone` + `.VAR` sequence, which will be much cheaper.