Add Any(iterable).isa-all(type) method
This returns a boolean whether all values produced by the iterable
match the given type.  Intended to be used in e.g. constraints such
as:
    sub foo(@a where .isa-all(Int))

as an alternative to:

    sub foo(@a where .all ~~ Int)

The .isa-all method is about 3.5x as fast as the .all ~~ Int approach
for an array consisting of 10 Ints.  The efficiency increase for non-
matching types will be *much* bigger, as is "isa-all" method can
short-circuit on any non-matching value, which the .all ~~ Int approach
can not.  So that would be *at least* 3.5x as fast, but potentially
orders of magnitude faster for bigger arrays.