New disp nativecall (#4629)
Lets callable objects have a say in which dispatcher is used when
they are invoked.

API for asking whether the compiler supports a certain nqp op.  This can be used to conditionally compile backend specific code in modules like NativeCall

Fix NativeCall performance regression on new-disp.  The trick of replacing the native routine's $!do attribute has never worked perfectly as closure cloning could lead to the generic code getting executed. For unknown reason that got a lot worse on new-disp.

Get around that by replacing the fragile $!do replacement with a dispatcher. Using the new CUSTOM-DISPATCHER hook, native routines get their own raku-nativecall dispatcher which for now just invokes the optimized body or the generic one, depending on availability.

This removes the equally fragile trick with generating function bodies in an exported CHECK phaser. To be replaced by dispatch programs at a later point.

Replace generic and specialized NativeCall code with boot-foreign-code dispatcher

Handle explicitly managed strings getting passed to native functions.  For explicitly managed strings, we need to pass the contained CStr value instead of the string object itself. This way NativeCall knows not to free the created C string after the call.

Note that the generated function bodies have never supported this. The problem just hasn't surfaced before, because the test runs the call only once and in the previous implementation the first call always went through the generic calling code (which does support explicitly managed strings).

Support passing Proxy to native routines.  For now if we find a Proxy in a native routine's arguments, just wrap the call in a simple function that fetches values from proxies as a side-effect.

Unbox Ints passed to native routines during dispatch.  Unboxing in HLL code can help spesh avoid boxing in the first place, thus speed up the calls. In addition, doing concreteness checks in HLL code can get rid of unnecessary branching.

Ignore fallback methods when looking for CUSTOM-DISPATCHER in dispatch.  When looking for a CUSTOM-DISPATCHER method on an object that is called, we want to ignore fallback methods (like FALLBACK or the junction autothreader), as those may give us an answer, even if there is no suitable CUSTOM-DISPATCHER method available, leading to very confusing errors.

Unbox CPointer objects when passed to native routines.  CPointer supports unboxing to int, so we should do that for the same reasons as for which we already do this with Ints.

Unbox strings passed to native routines, unbox nums for native calls.

Remove obsolete runtime compilation of native function bodies.  The raku-nativecall dispatcher handles everything these native function bodies were needed for in a better way. Note that this means that non-MoarVM backends use the generic nqp::nativecall op for all calls again, so this is a technically a performance regression there. The proper fix however is for them to support dispatchers.

Load raku-nativecall dispatcher code at BEGIN time.  We have to conditionally load this code as its using backend-specific NQP ops, but we can at least do so at BEGIN time instead of every time someone loads the NativeCall module.