Make Date.new(year,month,day) about 40% faster
By reducing the error checking in the method itself as little as
possible, and do the full awesome error generation in a separate
private method.  This brings !SET-SELF under the inlining limit,
thus making it 40% faster.  Alas, the subclass version did not
reach that limit yet, so not a lot to be gained there.