Fix CATCH blocks setting $! leading to serialization errors with BEGIN time exceptions
According to the docs $! only ought to be set by try blocks or statement
prefixes. CATCH blocks are supposed to only set $_ to the exception within
the block. Setting $! is not just wasteful but doing so at BEGIN time may
keep the exception referenced after being handled and prevent serialization.