Don't suggest method of the same name as not found
Instead of setting the score to 0, don't even add it to the suggestions.
If there are no other suggestions, before we could still use it, leading
to this example `my Int $a; $a.VAR.abs` giving this silly output:
`No such method 'abs' for invocant of type 'Scalar'.  Did you mean 'abs'?`
Now it just says `No such method 'abs' for invocant of type 'Scalar'`