Make the max number of threads depend on number of cores (#4652)
Previously, this was set to 64.  Clearly, should be dependent
on actual number of cores available.  For now, the maximum is set
to 8 times the number of cores, based on the previous value and the
more common 4-8 core hardware that was available when that maximum
was originally set, but make sure that we always have at least 64 threads
as before.