Fix segfaults due to concurrent hash access in parallel module loading
Read access to the %!seen and %!loaded hashes in CompUnit::Repository::
Installation and FileSystem was not protected by a lock. This could lead to us
fetching an only half-setup entry from the hash with a key that was still a
NULL pointer which we then dereferenced.

Fix by extending (or introducing) locks to all access to those hashes.