Add support for $RAKUDO_OPT environment variable
A way to specify default compiler CLI options, overridable by explicitly
specified in the command line. For example:

    RAKUDO_OPT="--I. --ll-exception" rakudo --Ilib test.raku

`--I.` would be ignored but `--ll-exeption` will be used.

Simple space split is used making it impossible to use a path with
a space in it to `--I`.

First `-e` or argument are ignored as well as any options after them,
similarly to what Rakudo compiler does.