Remove RESTRICTED setting (#3965)
RESTRICTED settings looks as if it is a security feature, but this
issue cannot be approached with a blacklist. For example, the things
that it blacklists are still available through method calls and
nativecall, and there is just no way to plug all the possible
holes. Moreover, those who actually need to run code securely will
probably resort to seccomp or the like, with a whitelist of allowed
syscalls, so the compiler seems to be the wrong place to implement
this.

I guess RESTRICTED setting was useful as a “don't accidentally break
the evalbot” feature, but it stopped working completely after the
release of v6.d.

This commit passes `test` (and `spectest`) as there seem to be no
tests at all for the feature.