Make FatRat literals codegen as WVals
A literal as .123.FatRat was codegenned as a Rat WVal and a call to
the .FatRat method at runtime: it is now codegenned as a FatRat WVal.

Also, if a Rat literal exceeds the precision of a Rat, it will issue
a worry that mentions to either make it a Num or a FatRat.