Bring Rational.floor/ceiling/round/Int below inlining limit
The error handling for divide-by-zero (which is an uncommon path to
take) made the methods too large to inline.  By moving this into a
separate private method, the common path becomes small enough to
not prevent inlining.

Did not benchmark, but it should have a positive effect on many
uses of these methods.