Introduce NYI as a sub
Many places in the core setting use:

   Failure.new: X::NYI.new: feature => "unimplemented feature"

in relatively hot code.  Another often occurring idiom is:

   X::NYI.new(feature => "unimplemented feature").throw

This sub allows these calls to be simplified to:

   NYI("unimplemented feature")        # saves 134 bytes

   NYI("unimplemented feature").throw  # saves 136 bytes

Since the feels like this could be useful for module developers
and/or general application developers as well, I decided to *not*
make this an implementation detail, but an actual feature of the
Raku Programming Language.

Please note that technically, the .throw version would unnecessarily
create a Failure object, but by the time we're going to throw the
exception, creating an extra object should hardly be an issue anyway.