Add missing precedence info for some infix ops (#4725)
This fixes the segfault uncovered by 71c62e7. Because `&infix:«+<»`
didn't have its precedence set, therefore the call to `.prec` went
through Code's `method prec`, which simple returns `my %`. For some
reason this is an HLL hash in the optimizer, so the atkey call in MoarVM
has problems. It shouldn't have segfaulted and probably still needs a
fix, but this change in Rakudo should happen regardless.