Add missing methods to Distribution::Resource
Distribution::Resource is supposed to be acting as an IO::Path object.
But because the underlying IO::Path object *SHOULD NOT EVER* be serialized
during pre-compilation, it is impossible to create an $.IO attribute.

Sadly, "handles" codegens to calling the attribute rather than any accessor
so we cannot create a dummy $.IO attribute and have all of the IO::Path
methods be "handled" by that (calling the accessor that is runtime only).

So manually add the "modified", "accessed", "changed" and "mode" methods.
And make the code a little more readable.  And add a comment to make sure
that nobody else in the future will try to create an $.IO atttribute.