Fix "This container does not reference a native unsigned integer"
When working with bufs and e.g. initializing them with lists of integers that
do not match in signedness, we need to give the runtime a chance to convert.
This happens automatically if the value is fetched into a scalar before passing
it to an nqp op. This is a bit less efficient than binding, but a sufficiently
advanced runtime can still optimize away the scalar container.