Remove unneeded/confusing can('prec') checks
This builds on commits caba0d3 and a7ccfc6 (admittedly not that
promptly!).  The prior commits moved the .prec method to Code with the
to prevent the need of checking whether various functions can prec;
this commit removes six such `can('prec')` checks in the `is equiv`,
`is tighter` and `is looser` traits.

Three of the six checks were literally unnecessary, because they were
on a Routine:D (which is guaranteed to have Code's prec method).  The
other 3 were on variables with a Callable (&) type constraint and thus
could technically have been necessary in the rare case that someone
passed a non-Code callable.  However, even in that edge case, the
check generated a confusing error message that suggested the trait
needed to be given an operator (in reality, any Code will do).

To address this edge case, this commit changes the parameter type
constraint from Callable to Code, which has the same effect but
produces a more accurate error message.