Decrement # of calls in a block if any are optimized away
This unblocks more removing of magic variables (`$¢`, `$!`, `$/`, `$_`),
which requires there to not be any calls in the block.

So this example `my int $a = 0; sub inc-a() { ++$a };` before would
have the first three magic variables still in the block, since even
though the call to `&prefix:<++>` was optimized away, the count was
still non-zero. Now they are all gone.