Optimize more cases of p6decontrv ops
If there's a QAST::Want in a QAST::Op(p6decontrv), try to optimize based on
the child of the QAST::Want. This turns the optimized version of `inc-a()` in
`my int $a = 0; sub inc-a() { ++$a };` into essentially the same as the
optimized version of `my int $a = 0; sub inc-a() { $a = $a + 1 };`,
i.e., `decont_i + assign_i + add_i` instead of `p6decontrv + assign_i +
add_i`.