Fix IO::Path::parent
The previous logic was too naive with absolute paths (just dropped the
basename) and messed up relative paths that had only updirs in the path,
but a basename.

The new implementation is now simpler by treating relative and absolute
paths the same and only looking at the basename to decide whether to append
an updir or drop an element.