Restore given/when optimization for Numeric cases
In the case of something like `given "a string" { when 5 { say "five" };
default { say "not five" } }`, the currently-not-working optimization
attempts to turn off `.Numeric` returning an (expensive to create) Failure
via `val` when we just need to know that it didn't succeed (i.e.,
`"a string".Numeric != 5`). However, just moving the `:fail-or-nil` argument
to the correct method call causes `Use of Nil in numeric context`. So
instead gen storing the result of the `.Numeric(:fail-or-nil)` in a
temporary variable and check whether it's `Nil`. If so, use `NaN`
instead of the given value, since that will never be == to the `when`
value (`when NaN` is explicitly handled earlier).

This test code:
```raku
my $a;
$a = do given "hi" {
    when 0 { "zero" };
    when 1 { "one" };
    when 2 { "two" };
    when 3 { "three" };
    when 4 { "four" };
    when 5 { "five" };
    when "hi" { "hello" };
    when * > 40 { "> 40" };
    default { "default" } } for ^10_000;
say now - INIT now;
say $a
```
drops from ~2s to ~0.25s.